+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Support for `Token Escrow` (XLS-85d)
 - Support for `NFTokenMintOffer` (XLS-52)
+- Add warning messages to `MPTokenIssuanceCreate` and `VaultCreate` transaction as per [XLS-89d](https://github.com/XRPLF/XRPL-Standards/tree/master/XLS-0089d-multi-purpose-token-metadata-schema).
+
 
 ## [4.2.0] - 2025-6-09